% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_em_mv.R
\name{run_em_algorithm}
\alias{run_em_algorithm}
\title{Run EM Algorithm for Imputation (Updating S3 Class)}
\usage{
run_em_algorithm(model, tolerance = 1e-05, max_iter = 100)
}
\arguments{
\item{model}{An object of class 'em_model'}

\item{tolerance}{Convergence tolerance for log-likelihood}

\item{max_iter}{Maximum number of iterations}
}
\value{
Updated 'em_model' object
}
\description{
Run EM Algorithm for Imputation (Updating S3 Class)
}
